Move plank templates into the config.
