+++ prow/Makefile
@@ -21,7 +21,7 @@ DECK_VERSION       = 0.30
 SPLICE_VERSION     = 0.20
 TOT_VERSION        = 0.2
 HOROLOGIUM_VERSION = 0.3
-PLANK_VERSION      = 0.23
+PLANK_VERSION      = 0.24
 
 # These are the usual GKE variables.
 PROJECT ?= k8s-prow
